[WIP] implement generation of lambda-layer types
