+++ plugin/deployment-configuration-maven-plugin/pom.xml
@@ -7,11 +7,11 @@
     <parent>
         <groupId>net.sf.aws-syndicate</groupId>
         <artifactId>deployment-configuration-processor</artifactId>
-        <version>1.5.5</version>
+        <version>1.5.6</version>
     </parent>
 
     <artifactId>deployment-configuration-maven-plugin</artifactId>
-    <version>1.5.5</version>
+    <version>1.5.6</version>
     <packaging>maven-plugin</packaging>
 
     <properties>
@@ -28,7 +28,7 @@
         <dependency>
             <groupId>net.sf.aws-syndicate</groupId>
             <artifactId>deployment-configuration-annotations</artifactId>
-            <version>1.5.5</version>
+            <version>1.5.6</version>
         </dependency>
         <dependency>
             <groupId>org.reflections</groupId>
@@ -108,6 +108,11 @@
             <groupId>io.github.openfeign</groupId>
             <artifactId>feign-jackson</artifactId>
             <version>10.0.1</version>
+        </dependency>
+        <dependency>
+            <groupId>net.sf.aws-syndicate</groupId>
+            <artifactId>deployment-configuration-maven-plugin</artifactId>
+            <version>1.5.0</version>
         </dependency>
     </dependencies>
