+++ plugin/deployment-configuration-maven-plugin/src/main/java/com/syndicate/deployment/api/model/response/SaveMetaResponse.java
@@ -1,4 +1,4 @@
-package com.syndicate.deployment.model.api.response;
+package com.syndicate.deployment.api.model.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
