+++ plugin/deployment-configuration-maven-plugin/src/main/java/com/syndicate/deployment/processor/IAnnotationProcessor.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2018 EPAM Systems, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.syndicate.deployment.processor;
+
+import org.apache.maven.plugin.MojoExecutionException;
+
+import java.util.Map;
+
+/**
+ * Created by Vladyslav Tereshchenko on 10/10/2016.
+ */
+public interface IAnnotationProcessor<T> {
+
+	Map<String, T> generateMeta(String absolutePath, String[] packages, String version, String fileName) throws MojoExecutionException;
+
+	Map<String, Object> convertMeta(Map<String, T> metaConfiguration);
+}
