+++ plugin/deployment-configuration-maven-plugin/src/main/java/com/syndicate/deployment/resolvers/impl/CliParametersCredentialResolver.java
@@ -1,6 +1,6 @@
 package com.syndicate.deployment.resolvers.impl;
 
-import com.syndicate.deployment.model.api.request.SyndicateCredentials;
+import com.syndicate.deployment.api.model.request.SyndicateCredentials;
 import com.syndicate.deployment.resolvers.AbstractChainedCredentialResolver;
 
 import java.security.InvalidParameterException;
